% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM_likelihood.R
\name{DM_negloglikelihood}
\alias{DM_negloglikelihood}
\title{Calculate the negative loglikelihood of the Dirichlet-Multinomial regression model.}
\usage{
DM_negloglikelihood(beta, X, Y)
}
\arguments{
\item{beta}{Matrix of \verb{(P+1) x J} parameters.}

\item{X}{Matrix of \verb{n x (P+1)} covariates.}

\item{Y}{Matrix of \verb{n x J} responses.}
}
\value{
Scalar value of the negative loglikelihood.
}
\description{
Calculate the negative loglikelihood of the Dirichlet-Multinomial regression model.
}
